Update README.md
The dataset on Indian Electoral Bonds offers a comprehensive glimpse into the intricate web of political funding in India. It unveils the identities of donors, their associated companies, and the recipient political parties, providing valuable insights into the dynamics of corporate interests within the political landscape. This dataset, while shedding light on crucial aspects of political financing, also underscores the ethical considerations surrounding corporate influence on democratic processes. It serves as a vital tool for researchers, policymakers, journalists, and civil society to delve into the complexities of political funding and advocate for transparency and accountability in governance.